Added semi-functionality to image scraper
On running, a piece of code will populate an array with the name of senators, as well as a link corresponding to their official photo. There is also a function that will take a name, and return the photo of this person. This function currently returns an image object, which is quite slow, however there is a commented line that will simply return the image URL instead. This function is called in order to add the image to the end of the "current" array, however I didn't touch any code regarding to the json, so it probably isn't properly stored.